+++ src/methods/format/_offset.js
@@ -1,15 +1,22 @@
 const fns = require('../../fns')
 
+// create the timezone offset part of an iso timestamp
+// it's kind of nuts how involved this is
 // "+01:00", "+0100", or simply "+01"
 const isoOffset = s => {
   let offset = s.timezone().current.offset
+  const isNegative = offset < 0
   let minute = '00'
-  if (offset % 1 === 0.5) {
-    //fraction of the hour
+  //handle 5.5 → '5:30'
+  if (Math.abs(offset % 1) === 0.5) {
     minute = '30'
-    offset = Math.floor(offset)
+    if (offset >= 0) {
+      offset = Math.floor(offset)
+    } else {
+      offset = Math.ceil(offset)
+    }
   }
-  if (offset < 0) {
+  if (isNegative) {
     //handle negative sign
     offset *= -1
     offset = fns.zeroPad(offset, 2)
@@ -19,7 +26,7 @@ const isoOffset = s => {
     offset = '+' + offset
   }
   offset = offset + ':' + minute
-  //this is a little cleaner?
+  //'Z' means 00
   if (offset === '+00:00') {
     offset = 'Z'
   }