+++ src/methods/query/normal.js
@@ -1,5 +1,4 @@
 const set = require('../set/set')
-const walkTo = require('../set/walk')
 
 //the most basic get/set methods
 let methods = {
@@ -32,11 +31,11 @@ let methods = {
     if (num !== undefined) {
       let s = this.clone()
       s.epoch = set.hours(s, num)
-      walkTo(s, {
-        month: this.month(),
-        date: this.date(),
-        hour: num
-      })
+      // walkTo(s, {
+      //   month: this.month(),
+      //   date: this.date(),
+      //   hour: num
+      // })
       return s
     }
     return d.getHours()